requirements.txt
pandas
numpy
scikit-learn
streamlit
matplotlib
seaborn
plotly
sqlalchemy
setup_mysql
joblib
jupyter
pymysql
cryptography